feat(ui): use new validateConnection
